fix bugs with config
