+++ custom_components/bedjet/climate.py
@@ -3,7 +3,7 @@ from datetime import datetime
 import logging
 import asyncio
 from homeassistant.components.climate import ClimateEntity
-from homeassistant.const import CONF_MAC, TEMP_FAHRENHEIT
+from homeassistant.const import TEMP_FAHRENHEIT
 from homeassistant.helpers.device_registry import format_mac
 from homeassistant.helpers.entity import DeviceInfo
 
@@ -31,41 +31,6 @@ from bleak import BleakClient, BleakError
 from bleak.backends.device import BLEDevice
 
 _LOGGER = logging.getLogger(__name__)
-
-
-
-async def async_setup_entry(hass, config_entry, async_add_entities):
-    mac = config_entry.data.get(CONF_MAC)
-    bedjets = await discover(hass)
-
-    if mac is not None:
-        bedjets = [bj for bj in bedjets if bj.mac == mac]
-
-    entities = []
-    for bj in bedjets:
-        my_device = BedjetDevice(bj)
-        entities.extend([BedjetDeviceEntity(my_device, idx) for idx in range(my_device.number_of_entities)])
-
-    async_add_entities(entities, True)
-
-async def async_setup_entry(hass, config_entry, async_add_entities):
-    mac = config_entry.data.get(CONF_MAC)
-    bedjets = await discover(hass)
-
-    # Filter devices based on MAC address, if applicable
-    if mac is not None:
-        bedjets = [bj for bj in bedjets if bj.mac == mac]
-
-    # Check if the list of discovered devices is empty
-    if not bedjets:
-        _LOGGER.warning("No BedJet devices were discovered.")
-        return
-
-    # Create BedjetDevice instance
-    bedjet_device = BedjetDevice(bedjets)
-
-    # Add entities to Home Assistant
-    async_add_entities(bedjet_device.entities, True)
 
 class BedjetDevice:
     def __init__(self, bedjets):
