+++ custom_components/bedjet/manifest.json
@@ -4,6 +4,6 @@
     "documentation": "https://github.com/asheliahut/ha-bedjet",
     "dependencies": [],
     "codeowners": ["@asheliahut"],
-    "version": "0.2.0",
+    "version": "0.2.1",
     "config_flow": true
 }
